refactor: split api service helpers by domain
