+++ cmd/api/units.go
@@ -74,6 +74,113 @@ func validateFloors(v *validator.Validator, floors []data.FloorCreate) {
 	}
 }
 
+func (app *application) prepareUnitForCreate(unit *data.Unit, input UnitCreate) ([]data.FloorCreate, error) {
+	if input.Type != "tower" {
+		return []data.FloorCreate{}, nil
+	}
+
+	towerData, err := parseTowerData(input.Data)
+	if err != nil {
+		return nil, err
+	}
+
+	unit.HousingUnitsCount = resolveHousingUnitsCount(input.HousingUnitsCount, towerData.HousingUnitsCount)
+	return towerData.Floors, nil
+}
+
+func (app *application) prepareUnitForUpdate(unit *data.Unit, input UnitUpdate) ([]data.FloorCreate, error) {
+	if unit.Type != "tower" {
+		return []data.FloorCreate{}, nil
+	}
+
+	if input.Data == nil {
+		unit.HousingUnitsCount = resolveHousingUnitsCount(input.HousingUnitsCount, unit.HousingUnitsCount)
+		return []data.FloorCreate{}, nil
+	}
+
+	towerData, err := parseTowerData(input.Data)
+	if err != nil {
+		return nil, err
+	}
+
+	unit.HousingUnitsCount = resolveHousingUnitsCount(input.HousingUnitsCount, towerData.HousingUnitsCount)
+	return towerData.Floors, nil
+}
+
+// insertUnit validates the unit and its floors, assigns a new UUID if needed, and persists it.
+func (app *application) insertUnit(unit *data.Unit, floors []data.FloorCreate) error {
+	v := validator.New()
+	data.ValidateUnit(v, unit)
+	if unit.Type != "tower" {
+		v.AddError("type", "invalid unit type")
+	}
+
+	if unit.Type == "tower" {
+		validateFloors(v, floors)
+		validateHousingUnitsCount(v, unit.HousingUnitsCount)
+	}
+
+	if !v.Valid() {
+		return &ValidationError{Errors: v.Errors}
+	}
+
+	if unit.ID == uuid.Nil {
+		unitID, err := uuid.NewV7()
+		if err != nil {
+			return err
+		}
+		unit.ID = unitID
+	}
+
+	err := app.models.Units.Insert(unit, floors)
+	if err != nil {
+		v := validator.New()
+		switch {
+		case errors.Is(err, data.ErrDuplicateFloorIndexes):
+			v.AddError("floors", "floor indexes must be unique")
+			return &ValidationError{Errors: v.Errors}
+		case errors.Is(err, data.ErrFloorIndexGap):
+			v.AddError("floors", "floor indexes must be continuous without gaps")
+			return &ValidationError{Errors: v.Errors}
+		default:
+			return err
+		}
+	}
+
+	return nil
+}
+
+func (app *application) updateUnit(unit *data.Unit, floors []data.FloorCreate) error {
+	v := validator.New()
+	data.ValidateUnit(v, unit)
+
+	if unit.Type == "tower" {
+		validateFloors(v, floors)
+		validateHousingUnitsCount(v, unit.HousingUnitsCount)
+	}
+
+	if !v.Valid() {
+		return &ValidationError{Errors: v.Errors}
+	}
+
+	err := app.models.Units.Update(unit, floors)
+	if err != nil {
+		v := validator.New()
+		switch {
+		case errors.Is(err, data.ErrDuplicateFloorIndexes):
+			v.AddError("floors", "floor indexes must be unique")
+			return &ValidationError{Errors: v.Errors}
+		case errors.Is(err, data.ErrFloorIndexGap):
+			v.AddError("floors", "floor indexes must be continuous without gaps")
+			return &ValidationError{Errors: v.Errors}
+		default:
+			return err
+		}
+	}
+
+	return nil
+}
+
 func (app *application) createUnitHandler(w http.ResponseWriter, r *http.Request) {
 	projectID, err := app.readUUIDParam(r, "projectID")
 	if err != nil {
